refactor: applyMoviesToDailyBoxOfficeItem 메서드 구현, 뷰컨트롤러의 loadImage 메서드를 NetworkManager 메서드로 변경
