+++ BoxOffice/Model/Network/NetworkManager.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import UIKit
 
 struct NetworkManager {
     let urlSession: URLSession
@@ -38,6 +39,36 @@ struct NetworkManager {
             
             if let data = data {
                 guard let result = Decoder.parseJSON(data, returnType: returnType) else {
+                    completion(.failure(.decode))
+                    return
+                }
+                completion(.success(result))
+            }
+        }
+        task.resume()
+    }
+    
+    func loadImage(from imageURL: URL, completion: @escaping ((Result<UIImage, NetworkError>) -> Void)) {
+        let urlRequest = URLRequest(url: imageURL)
+        
+        let task = URLSession.shared.dataTask(with: urlRequest) { data, response, error in
+            if error != nil {
+                completion(.failure(.unknown))
+                return
+            }
+            
+            guard let httpResponse = response as? HTTPURLResponse else {
+                completion(.failure(.httpResponse))
+                return
+            }
+            
+            guard (200...299).contains(httpResponse.statusCode) else {
+                completion(.failure(.httpStatusCode(code: httpResponse.statusCode)))
+                return
+            }
+            
+            if let data = data {
+                guard let result = UIImage(data: data) else {
                     completion(.failure(.decode))
                     return
                 }
